fix duplicated CORS error

+++ core-banking/src/main/java/com/example/corebanking/account/dto/AccountCreateRequest.java
@@ -3,10 +3,12 @@ package com.example.corebanking.account.dto;
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotNull;
 
+import java.math.BigDecimal;
+
 public record AccountCreateRequest(
         @NotNull(message = "User Id is required")
         Long userId,
 
         @Min(value = 1000, message = "A minimum deposit of 1,000 yen is required to create an account.")
-        Long initialBalance
+        BigDecimal initialBalance
 ) {}
